refactor: Request 파싱 로직 각 객체에게 분리 및 상수 분리
