+++ tomcat/src/test/java/support/RequestFixture.java
@@ -1,7 +1,7 @@
 package support;
 
-import static org.apache.coyote.http11.request.Request.HEADER_KEY_VALUE_DELIMITER;
 import static org.apache.coyote.http11.request.Request.SPACE_DELIMITER;
+import static org.apache.coyote.http11.request.RequestHeaders.HEADER_KEY_VALUE_DELIMITER;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
